@@ -11,31 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.1",
+    "@angular/animations": "^8.2.13",
     "@angular/cdk": "^6.4.7",
-    "@angular/common": "^7.0.1",
-    "@angular/compiler": "^7.0.1",
-    "@angular/core": "^7.0.1",
+    "@angular/common": "^8.2.13",
+    "@angular/compiler": "^8.2.13",
+    "@angular/core": "^8.2.13",
     "@angular/flex-layout": "git+https://github.com/angular/flex-layout-builds.git",
-    "@angular/forms": "^7.0.1",
+    "@angular/forms": "^8.2.13",
     "@angular/http": "^7.0.1",
     "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^7.0.1",
-    "@angular/platform-browser-dynamic": "^7.0.1",
-    "@angular/router": "^7.0.1",
+    "@angular/platform-browser": "^8.2.13",
+    "@angular/platform-browser-dynamic": "^8.2.13",
+    "@angular/router": "^8.2.13",
     "angular2-csv": "^0.2.9",
     "chart.js": "^2.7.3",
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "ng2-charts": "^1.6.0",
-    "rxjs": "^6.3.3",
-    "zone.js": "~0.8.26"
+    "rxjs": "^6.5.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.0",
-    "@angular/cli": "~7.0.3",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/language-service": "^7.0.1",
+    "@angular-devkit/build-angular": "~0.803.17",
+    "@angular/cli": "~8.3.17",
+    "@angular/compiler-cli": "^8.2.13",
+    "@angular/language-service": "^8.2.13",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -50,6 +50,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.1.3"
+    "typescript": "^3.5.3"
   }
 }
